feat: testing quickjs sanbox
