+++ packages/bruno-js/.gitignore
@@ -1,1 +1,1 @@
-src/bundle-browser-rollup.js
+src/sandbox/bundle-browser-rollup.js
