+++ packages/bruno-js/package.json
@@ -12,11 +12,10 @@
   },
   "scripts": {
     "test": "jest --testPathIgnorePatterns test.js",
-    "postinstall": "npm run isolated-vm:install && npm run isolated-vm:bundle-libraries",
+    "sandbox:bundle-libraries": "node ./src/sandbox/bundle-libraries.js",
     "isolated-vm:install": "cd ./node_modules/isolated-vm && npm install",
     "isolated-vm:prebuild:dev": "node ./scripts/prebuild-isolated-vm-for-dev.js",
-    "isolated-vm:prebuild:prod": "node ./scripts/prebuild-isolated-vm-for-prod-builds.js",
-    "isolated-vm:bundle-libraries": "node ./src/sandbox/isolatedvm/utils/bundle-libraries.js"
+    "isolated-vm:prebuild:prod": "node ./scripts/prebuild-isolated-vm-for-prod-builds.js"
   },
   "dependencies": {
     "@usebruno/common": "0.1.0",
