--- packages/bruno-js/src/sandbox/isolatedvm copy/shims/lib/nanoid.js
@@ -1,23 +1,0 @@
-const ivm = require('isolated-vm');
-const { nanoid } = require('nanoid');
-
-const addNanoidShimToContext = async (context) => {
-  await context.evalClosure(
-    `
-        globalThis.nanoid = {};
-        globalThis.nanoid.nanoid = () => $0.applySync(undefined);
-        globalThis.requireObject = {
-            ...globalThis.requireObject,
-            nanoid: globalThis.nanoid
-        }
-    `,
-    [
-      () => {
-        return new ivm.ExternalCopy(nanoid()).copyInto({ release: true });
-      }
-    ],
-    { arguments: { reference: true } }
-  );
-};
-
-module.exports = addNanoidShimToContext;
