+++ packages/bruno-js/src/sandbox/quickjs/utils/index.js
@@ -3,6 +3,10 @@ const marshallToVm = (value, vm) => {
     return vm.undefined;
   }
 
+  if (value === null) {
+    return vm.null;
+  }
+
   if (typeof value === "string") {
     return vm.newString(value);
   } else if (typeof value === "number") {
@@ -13,13 +17,13 @@ const marshallToVm = (value, vm) => {
     if (Array.isArray(value)) {
       const arr = vm.newArray();
       for (let i = 0; i < value.length; i++) {
-        vm.setElement(arr, i, marshallToVm(value[i]));
+        vm.setProp(arr, i, marshallToVm(value[i], vm));
       }
       return arr;
     } else {
       const obj = vm.newObject();
       for (const key in value) {
-        vm.setProp(obj, key, marshallToVm(value[key]));
+        vm.setProp(obj, key, marshallToVm(value[key], vm));
       }
       return obj;
     }
