+++ packages/bruno-tests/collection/scripting/api/req/getUrl.bru
@@ -10,7 +10,6 @@ get {
   auth: none
 }
 
-
 assert {
   res.status: eq 200
   res.body: eq pong
